Add login fixture and refactor auth tests to use it
