+++ e2e/config/fixtures.js
@@ -2,7 +2,10 @@ import { test as base } from '@playwright/test'
 import fs from 'fs'
 import path from 'path'
 
-import { EmailService } from '../utils'
+import { HttpStatus } from '../../src/lib/net'
+import { LOGIN_PATH } from '../../src/services/backend/paths'
+import { fillLoginFormAndSubmit } from '../scenarios'
+import { EmailService, waitForApiCall } from '../utils'
 
 const LOCALES_PATH = './public/locales'
 
@@ -14,7 +17,25 @@ const loadTranslations = (locale = 'en') => {
 
 export const test = base.extend({
   t: [loadTranslations(), { scope: 'worker' }],
-  emailService: [new EmailService(), { scope: 'worker' }]
+
+  emailService: [new EmailService(), { scope: 'worker' }],
+
+  login: async ({ page, t }, use) => {
+    const login = async credentials => {
+      await page.goto('/login')
+
+      const apiPromise = waitForApiCall(page, {
+        path: LOGIN_PATH,
+        status: HttpStatus.OK
+      })
+
+      await fillLoginFormAndSubmit(page, credentials, t)
+      await apiPromise
+    }
+
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    await use(login)
+  }
 })
 
 export { expect } from '@playwright/test'
